extract stream buffering behavior into submodule
